+++ examples/example3.cc
@@ -15,7 +15,7 @@ int main()
 
   fftw::maxthreads=get_max_threads();
 
-  unsigned int nx=4, ny=4, nz=4;
+  size_t nx=4, ny=4, nz=4;
   size_t align=sizeof(Complex);
 
   array3<Complex> f(nx,ny,nz,align);
@@ -23,9 +23,9 @@ int main()
   fft3d Forward(-1,f);
   fft3d Backward(1,f);
 
-  for(unsigned int i=0; i < nx; i++)
-    for(unsigned int j=0; j < ny; j++)
-      for(unsigned int k=0; k < nz; k++)
+  for(size_t i=0; i < nx; i++)
+    for(size_t j=0; j < ny; j++)
+      for(size_t k=0; k < nz; k++)
         f(i,j,k)=Complex(10*k+i,j);
 
   cout << "\ninput:\n" << f;
