+++ examples/examplecconv.cc
@@ -10,9 +10,9 @@ using namespace utils;
 using namespace fftwpp;
 
 
-void init(Complex *f, Complex *g, unsigned int m)
+void init(Complex *f, Complex *g, size_t m)
 {
-  for(unsigned int k=0; k < m; k++) {
+  for(size_t k=0; k < m; k++) {
     f[k]=Complex(k,k+1);
     g[k]=Complex(k,2*k+1);
   }
@@ -23,7 +23,7 @@ int main(int argc, char* argv[])
   fftw::maxthreads=get_max_threads();
 
   // size of problem
-  unsigned int m=8;
+  size_t m=8;
 
 #ifndef __SSE2__
   fftw::effort |= FFTW_NO_SIMD;
@@ -36,7 +36,7 @@ int main(int argc, char* argv[])
   cout << "1d non-centered complex convolution:" << endl;
   init(f,g,m);
   cout << "\ninput:\nf\tg" << endl;
-  for(unsigned int i=0; i < m; i++)
+  for(size_t i=0; i < m; i++)
     cout << f[i] << "\t" << g[i] << endl;
 
   ImplicitConvolution C(m);
@@ -44,7 +44,7 @@ int main(int argc, char* argv[])
   C.convolve(f,g);
 
   cout << "\noutput:" << endl;
-  for(unsigned int i=0; i < m; i++) cout << f[i] << endl;
+  for(size_t i=0; i < m; i++) cout << f[i] << endl;
 
   deleteAlign(g);
   deleteAlign(f);
