+++ examples/examplecconv3.cc
@@ -11,21 +11,21 @@ using namespace Array;
 using namespace fftwpp;
 
 // size of problem
-unsigned int mx=4;
-unsigned int my=4;
-unsigned int mz=4;
+size_t mx=4;
+size_t my=4;
+size_t mz=4;
 
-inline void init(array3<Complex>& f, array3<Complex>& g, unsigned int M=1)
+inline void init(array3<Complex>& f, array3<Complex>& g, size_t M=1)
 {
   double factor=1.0/sqrt((double) M);
-  for(unsigned int s=0; s < M; ++s) {
+  for(size_t s=0; s < M; ++s) {
     double S=sqrt(1.0+s);
     double ffactor=S*factor;
     double gfactor=1.0/S*factor;
-    for(unsigned int i=0; i < mx; ++i) {
-      unsigned int I=s*mx+i;
-      for(unsigned int j=0; j < my; j++) {
-        for(unsigned int k=0; k < mz; k++) {
+    for(size_t i=0; i < mx; ++i) {
+      size_t I=s*mx+i;
+      for(size_t j=0; j < my; j++) {
+        for(size_t k=0; k < mz; k++) {
           f[I][j][k]=ffactor*Complex(i+k,j+k);
           g[I][j][k]=gfactor*Complex(2*i+k,j+1+k);
         }
