+++ examples/exampletranspose.cc
@@ -14,7 +14,7 @@ int main()
   cout << "Tranposition of complex variables using the Array class" << endl;
   fftw::maxthreads=get_max_threads();
 
-  unsigned int nx=4, ny=4;
+  size_t nx=4, ny=4;
   size_t align=sizeof(Complex);
 
   array2<Complex> f(nx,ny,align);
@@ -23,8 +23,8 @@ int main()
   Transpose T(nx,ny,1,f(),ft(),fftw::maxthreads);
   Transpose Tinv(ny,nx,1,ft(),f(),fftw::maxthreads);
 
-  for(unsigned int i=0; i < nx; i++)
-    for(unsigned int j=0; j < ny; j++)
+  for(size_t i=0; i < nx; i++)
+    for(size_t j=0; j < ny; j++)
       f(i,j)=Complex(i,j);
 
   cout << "\ninput:\n" << f;
